slider goes up to 28 days now

+++ stop_services.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-
-tmux kill-ses -t temperature_logger
-
-echo "Temperature logger services killed."
+if tmux has-session -t temperature_logger; then
+    tmux kill-ses -t temperature_logger
+    echo "Temperature logger services killed."
+else
+    echo "Temperature logging services are not running."
+fi
